The following exception has been fixed. Message: 	Failed to connect to an IPC Port: The system cannot find the file specified. Full Trace: 	System.Runtime.Remoting.RemotingException: Failed to connect to an IPC Port: The system cannot find the file specified. Etc...
Details:
This can be generated when a user spam clicks on the "download with manager" link, or attempts to download multiple mods with in a very short amount of time.
What happens is as I suspected, the second (or third etc...) instance of NMM attempts to access the messenger before the first instance has finished initialising, but has claimed the mutex. This goes wrong and an exception is thrown.

I have fixed this by building in a redundancy so that the messenger will attempt to be accessed three times over fifteen seconds, which should hopefully allow the first instance of NMM to create the messenger.

--HG--
branch : thomaslee-bugfix-ipc